Tiny fixes

+++ midiplayer/new/MIDIPlayer.h
@@ -40,7 +40,7 @@ enum
     default_cfg_xmiloopz = 1,
     default_cfg_ff7loopz = 1,
 
-    DefaultMIDIFlavor = MIDIFlavor::None,
+    DefaultMIDIFlavor = (int)MIDIFlavor::None,
     DefaultUseMIDIEffects = 1,
     DefaultUseSuperMuntWithMT32 = 1,
     DefaultUseSecretSauceWithXG = 0,
